inference-model/requirements.txt
numpy==1.26.4
pillow==10.3.0
tensorflow==2.15.0
fastapi==0.111.0
uvicorn==0.30.1